remove inline mock of orderbook
